refactor - refactoring category api
