+++ constants/react-query/index.ts
@@ -1,4 +1,7 @@
 export const queryKey = {
+  category: (isExcluded: boolean) => {
+    return [isExcluded ? 'excludedCategory' : 'category'];
+  },
   productItemList: (requestParams: Omit<req.ShopFeed, 'page' | 'size'>) => {
     return ['productItemList', { ...requestParams }];
   },
